refactor(saga): collapse single-consumer test_support fixture into saga.rs + drop dead payment_adapter param
Two pure test-code dedup cleanups following the escrow-correction in
ce57bdd0d (which passed `None` for the payment adapter, leaving two stale
artifacts):

1. `VoidCountingPaymentAdapter` now has exactly one consumer (the actor
   saga-handler tests in saga.rs). The "shared across modules" rationale that
   justified hoisting it into a `pub(crate) mod test_support` is now false, so
   inline the struct + impl back into saga.rs's `#[cfg(test)] mod tests` as a
   private fixture, delete test_support.rs, and drop its `mod` declaration in
   context/mod.rs. Doc-comment loses the now-false "shared by ..." line.

2. The `xctx_supervisor_with_journal` test helper took a
   `payment_adapter: Option<Arc<dyn PaymentAdapterDyn>>` that all three
   callsites passed `None` for. Drop the dead parameter and forward `None`
   from a single internal site with a comment explaining why the live §6.2.4
   xctx outbound path stages no caller-side escrow.

No production logic changes; behavior is identical.
